pyproject.toml
[tool.poetry]
name = "vocab-builder"
version = "0.1.0"
description = "A project to help build vocabulary efficiently."
authors = ["Robin <robin@vespertool.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "*", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pandas = "^2.2.3"
sshtunnel = "^0.4.0"
pymysql = "^1.1.1"
sqlalchemy = "^2.0.36"
pyyaml = "^6.0.2"
gitpython = "^3.1.43"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
requests = "^2.32.3"
openai = "^1.55.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
